requirements.txt
pyodbc
sqlalchemy
langchain
langchain-community
faiss-cpu
sentence-transformers
openai
chromadb
pydantic
python-dotenv